# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinAI CA API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubmitSurveyBody {
  name: string;
  age: string;
  gender: string;
  qualification: string;
  /** @nullable */
  qualificationOther?: string | null;
  services: string[];
  /** @nullable */
  experience?: string | null;
  ideas: string;
}

export interface SurveyResponse {
  id: number;
  name: string;
  createdAt: string;
}

export interface SubjectScore {
  subject: string;
  /** @nullable */
  score?: number | null;
  maxScore: number;
}

export interface AnalyzeDiagnosticBody {
  name: string;
  email: string;
  examLevel: string;
  attemptNumber: number;
  subjects: SubjectScore[];
  weakAreas: string[];
  /** @nullable */
  studyHours?: string | null;
}

export interface DiagnosticResponse {
  id: number;
  name: string;
  email: string;
  examLevel: string;
  attemptNumber: number;
  subjectsJson: string;
  weakAreas: string[];
  /** @nullable */
  studyHours?: string | null;
  /** @nullable */
  report?: string | null;
  createdAt: string;
}