# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinAI CA API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit survey response
 */
export const SubmitSurveyBody = zod.object({
  name: zod.string(),
  age: zod.string(),
  gender: zod.string(),
  qualification: zod.string(),
  qualificationOther: zod.string().nullish(),
  services: zod.array(zod.string()),
  experience: zod.string().nullish(),
  ideas: zod.string(),
});

/**
 * @summary Submit CA exam data for AI diagnostic analysis (streams SSE)
 */
export const AnalyzeDiagnosticBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  examLevel: zod.string(),
  attemptNumber: zod.number(),
  subjects: zod.array(
    zod.object({
      subject: zod.string(),
      score: zod.number().nullish(),
      maxScore: zod.number(),
    }),
  ),
  weakAreas: zod.array(zod.string()),
  studyHours: zod.string().nullish(),
});

/**
 * @summary Get a saved diagnostic report
 */
export const GetDiagnosticParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDiagnosticResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  examLevel: zod.string(),
  attemptNumber: zod.number(),
  subjectsJson: zod.string(),
  weakAreas: zod.array(zod.string()),
  studyHours: zod.string().nullish(),
  report: zod.string().nullish(),
  createdAt: zod.string(),
});