# lib/api-zod/src/generated/types/analyzeDiagnosticBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinAI CA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectScore } from "./subjectScore";

export interface AnalyzeDiagnosticBody {
  name: string;
  email: string;
  examLevel: string;
  attemptNumber: number;
  subjects: SubjectScore[];
  weakAreas: string[];
  /** @nullable */
  studyHours?: string | null;
}