# lib/api-zod/src/generated/types/diagnosticResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinAI CA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiagnosticResponse {
  id: number;
  name: string;
  email: string;
  examLevel: string;
  attemptNumber: number;
  subjectsJson: string;
  weakAreas: string[];
  /** @nullable */
  studyHours?: string | null;
  /** @nullable */
  report?: string | null;
  createdAt: string;
}