# lib/api-zod/src/generated/types/submitSurveyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinAI CA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitSurveyBody {
  name: string;
  age: string;
  gender: string;
  qualification: string;
  /** @nullable */
  qualificationOther?: string | null;
  services: string[];
  /** @nullable */
  experience?: string | null;
  ideas: string;
}