# lib/api-zod/src/generated/types/surveyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinAI CA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SurveyResponse {
  id: number;
  name: string;
  createdAt: string;
}